+++ WebServer/WebServer.py
@@ -24,7 +24,7 @@ from urllib.parse import urlparse, urlsplit, urldefrag, parse_qs
 from time import time, ctime, strftime, gmtime, mktime, strptime
 
 from Modules.zigateConsts import ADDRESS_MODE, MAX_LOAD_ZIGATE, ZCL_CLUSTERS_LIST , CERTIFICATION_CODE, PROFILE_ID, ZHA_DEVICES, ZLL_DEVICES, ZIGATE_COMMANDS, ZCL_CLUSTERS_ACT
-from Modules.bindings import webBind, webUnBind
+
 from Modules.basicOutputs import ZigatePermitToJoin, sendZigateCmd, start_Zigate, setExtendedPANID, zigateBlueLed
 from Modules.legrand_netatmo import legrand_ledInDark, legrand_ledIfOnOnOff, legrand_dimOnOff, legrand_ledShutter
 from Modules.actuators import actuators
@@ -32,11 +32,6 @@ from Modules.tools import is_hex
 from Classes.PluginConf import PluginConf,SETTINGS
 from Classes.GroupMgt import GroupsManagement
 from Classes.DomoticzDB import DomoticzDB_Preferences
-
-
-
-
-
 
 MIMETYPES = { 
         "gif": "image/gif" ,
@@ -64,20 +59,20 @@ MIMETYPES = {
         "woff": "application/x-font-woff" 
    }
 
-
-
 class WebServer(object):
 
-    from WebServer.WebServerLogging import logging
-    from WebServer.WebServersession import startWebServer, onStop, onConnect, onDisconnect
+    from WebServer.logging import logging
+    from WebServer.com import startWebServer, onStop, onConnect, onDisconnect
 
     from WebServer.onMessage import onMessage
     from WebServer.sendresponse import sendResponse
 
     from WebServer.restDispatch import do_rest
+
     from WebServer.tools import setupHeadersResponse
 
     hearbeats = 0 
+
     def __init__( self, networkenergy, networkmap, ZigateData, PluginParameters, PluginConf, Statistics, adminWidgets, ZigateComm, HomeDirectory, hardwareID, DevicesInPairingMode, groupManagement, Devices, ListOfDevices, IEEE2NWK , permitTojoin, WebUserName, WebPassword, PluginHealth, httpPort, loggingFileHandle):
 
         self.httpServerConn = None
@@ -119,82 +114,6 @@ class WebServer(object):
         # Start the WebServer
         self.startWebServer( )                    
 
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
     def rest_plugin_health( self, verb, data, parameters):
 
         _response = setupHeadersResponse()
@@ -207,12 +126,14 @@ class WebServer(object):
             _response["Headers"]["Pragma"] = "no-cache"
             _response["Headers"]["Expires"] = "0"
             _response["Headers"]["Accept"] = "*/*"
+
         _response["Status"] = "200 OK"
         _response["Headers"]["Content-Type"] = "application/json; charset=utf-8"
         if verb == 'GET':
-            health = {}
-            health['HealthFlag'] = self.PluginHealth['Flag']
-            health['HealthTxt'] = self.PluginHealth['Txt']
+            health = {
+                'HealthFlag': self.PluginHealth['Flag'],
+                'HealthTxt': self.PluginHealth['Txt'], }
+
             if 'Firmware Update' in self.PluginHealth:
                 health['OTAupdateProgress'] = self.PluginHealth['Firmware Update']['Progress']
                 health['OTAupdateDevice'] = self.PluginHealth['Firmware Update']['Device']
@@ -222,101 +143,8 @@ class WebServer(object):
 
             _response["Data"] = json.dumps( health, sort_keys=True )
 
-
         return _response
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-    def rest_req_nwk_inter( self, verb, data, parameters):
-
-        _response = setupHeadersResponse()
-        if self.pluginconf.pluginConf['enableKeepalive']:
-            _response["Headers"]["Connection"] = "Keep-alive"
-        else:
-            _response["Headers"]["Connection"] = "Close"
-        if not self.pluginconf.pluginConf['enableCache']:
-            _response["Headers"]["Cache-Control"] = "no-cache, no-store, must-revalidate"
-            _response["Headers"]["Pragma"] = "no-cache"
-            _response["Headers"]["Expires"] = "0"
-            _response["Headers"]["Accept"] = "*/*"
-        _response["Status"] = "200 OK"
-        _response["Headers"]["Content-Type"] = "application/json; charset=utf-8"
-        if verb == 'GET':
-            action = {}
-            action['Name'] = "Nwk-Interferences"
-            action['TimeStamp'] = int(time())
-            _response["Data"] = json.dumps( action, sort_keys=True )
-
-            if self.pluginparameters['Mode1'] != 'None':
-                if self.networkenergy:
-                    self.networkenergy.start_scan()
-
-        return _response
-
-    def rest_req_nwk_full( self, verb, data, parameters):
-
-        _response = setupHeadersResponse()
-        if self.pluginconf.pluginConf['enableKeepalive']:
-            _response["Headers"]["Connection"] = "Keep-alive"
-        else:
-            _response["Headers"]["Connection"] = "Close"
-        if not self.pluginconf.pluginConf['enableCache']:
-            _response["Headers"]["Cache-Control"] = "no-cache, no-store, must-revalidate"
-            _response["Headers"]["Pragma"] = "no-cache"
-            _response["Headers"]["Expires"] = "0"
-            _response["Headers"]["Accept"] = "*/*"
-
-        _response["Status"] = "200 OK"
-        _response["Headers"]["Content-Type"] = "application/json; charset=utf-8"
-        if verb == 'GET':
-            action = {}
-            action['Name'] = "Nwk-Energy-Full"
-            action['TimeStamp'] = int(time())
-            _response["Data"] = json.dumps( action, sort_keys=True )
-
-            if self.pluginparameters['Mode1'] != 'None':
-                if self.networkenergy:
-                    self.networkenergy.start_scan( root='0000', target='0000')
-
-        return _response
-
-    def rest_req_topologie( self, verb, data, parameters):
-        _response = setupHeadersResponse()
-        if self.pluginconf.pluginConf['enableKeepalive']:
-            _response["Headers"]["Connection"] = "Keep-alive"
-        else:
-            _response["Headers"]["Connection"] = "Close"
-        _response["Status"] = "200 OK"
-        _response["Headers"]["Content-Type"] = "application/json; charset=utf-8"
-        if verb == 'GET':
-            action = {}
-            action['Name'] = 'Req-Topology'
-            action['TimeStamp'] = int(time())
-            _response["Data"] = json.dumps( action, sort_keys=True )
-
-            self.logging( 'Log', "Request a Start of Network Topology scan")
-            if self.networkmap:
-                if not self.networkmap.NetworkMapPhase():
-                    self.networkmap.start_scan()
-                else:
-                    self.logging( 'Log', "Cannot start Network Topology as one is in progress...")
-        return _response
-
+                    
     def rest_zigate_erase_PDM( self, verb, data, parameters):
 
         _response = setupHeadersResponse()
@@ -338,10 +166,9 @@ class WebServer(object):
                 self.logging( 'Status', "ZigateConf - Setting extPANID : 0x%016x" %( self.pluginconf.pluginConf['extendedPANID'] ))
                 if self.pluginparameters['Mode1'] != 'None':
                     setExtendedPANID(self, self.pluginconf.pluginConf['extendedPANID'])
-            action = {}
-            action['Description'] = 'Erase Zigate PDM - Non Implemente'
-            #if self.pluginparameters['Mode1'] != 'None':
-            #    start_Zigate( self )
+            action = {'Description': 'Erase Zigate PDM - Non Implemente'}
+                #if self.pluginparameters['Mode1'] != 'None':
+                #    start_Zigate( self )
         return _response
 
     def rest_rescan_group( self, verb, data, parameters):
@@ -404,13 +231,14 @@ class WebServer(object):
             if self.zigatedata:
                 _response["Data"] = json.dumps( self.zigatedata, sort_keys=True )
             else:
-                fake_zigate = {}
-                fake_zigate['Firmware Version'] = "fake - 0310"
-                fake_zigate['IEEE'] = "00158d0001ededde"
-                fake_zigate['Short Address'] = "0000"
-                fake_zigate['Channel'] = "0b"
-                fake_zigate['PANID'] = "51cf"
-                fake_zigate['Extended PANID'] = "bd1247ec9d358634"
+                fake_zigate = {
+                    'Firmware Version': 'fake - 0310',
+                    'IEEE': '00158d0001ededde',
+                    'Short Address': '0000',
+                    'Channel': '0b',
+                    'PANID': '51cf',
+                    'Extended PANID': 'bd1247ec9d358634',
+                }
 
                 _response["Data"] = json.dumps( fake_zigate , sort_keys=True )
         return _response
@@ -448,198 +276,6 @@ class WebServer(object):
         _response["Headers"]["Content-Type"] = "application/json; charset=utf-8"
         if verb == 'GET':
                 _response["Data"] = json.dumps( self.pluginparameters, sort_keys=True )
-        return _response
-
-    def rest_netTopologie( self, verb, data, parameters):
-
-        _filename = self.pluginconf.pluginConf['pluginReports'] + 'NetworkTopology-v3-' + '%02d' %self.hardwareID + '.json'
-        self.logging( 'Debug', "Filename: %s" %_filename)
-
-        _response = setupHeadersResponse()
-        if self.pluginconf.pluginConf['enableKeepalive']:
-            _response["Headers"]["Connection"] = "Keep-alive"
-        else:
-            _response["Headers"]["Connection"] = "Close"
-        _response["Data"] = "{}"
-        _response["Status"] = "200 OK"
-        _response["Headers"]["Content-Type"] = "application/json; charset=utf-8"
-
-        if not os.path.isfile( _filename ) :
-            _response['Data'] = json.dumps( {} , sort_keys=True ) 
-            return _response
-
-        # Read the file, as we have anyway to do it
-        _topo = {}           # All Topo reports
-        _timestamps_lst = [] # Just the list of Timestamps
-        with open( _filename , 'rt') as handle:
-            for line in handle:
-                if line[0] != '{' and line[-1] != '}': 
-                    continue
-                
-                entry = json.loads( line, encoding=dict )
-                for _ts in entry:
-                    _timestamps_lst.append( int(_ts) )
-                    _topo[_ts] = [] # List of Father -> Child relation for one TimeStamp
-                    _check_duplicate = []
-                    _nwkid_list = []
-                    reportLQI = entry[_ts]
-
-                    for item in reportLQI:
-                        self.logging( 'Debug', "Node: %s" %item)
-                        if item != '0000' and item not in self.ListOfDevices:
-                            continue
-
-                        if item not in _nwkid_list:
-                            _nwkid_list.append( item )
-                        for x in  reportLQI[item]['Neighbours']:
-                            self.logging( 'Debug', "---> %s" %x)
-                            # Report only Child relationship
-                            if x != '0000' and x not in self.ListOfDevices: 
-                                continue
-
-                            if item == x: 
-                                continue
-
-                            if 'Neighbours' not in reportLQI[item]:
-                                Domoticz.Error("Missing attribute :%s for (%s,%s)" %('Neighbours', item, x))
-                                continue
-
-                            for attribute in ( '_relationshp', '_lnkqty', '_devicetype', '_depth' ):
-                                if attribute not in reportLQI[item]['Neighbours'][x]:
-                                    Domoticz.Error("Missing attribute :%s for (%s,%s)" %(attribute, item, x))
-                                    continue
-
-                            if x not in _nwkid_list:
-                                _nwkid_list.append( x )
-                            
-                            # We need to reorganise in Father/Child relationship.
-                            if reportLQI[item]['Neighbours'][x]['_relationshp'] == 'Parent':
-                                _father = item
-                                _child  = x
-
-                            elif reportLQI[item]['Neighbours'][x]['_relationshp'] == 'Child':
-                                _father = x
-                                _child = item
-
-                            elif reportLQI[item]['Neighbours'][x]['_relationshp'] == 'Sibling':
-                                _father = item
-                                _child  = x
-
-                            elif reportLQI[item]['Neighbours'][x]['_relationshp'] == 'Former Child':
-                                # Not a Parent, not a Child, not a Sibbling
-                                #_father = item
-                                #_child  = x
-                                continue
-
-                            elif reportLQI[item]['Neighbours'][x]['_relationshp'] == 'None':
-                                # Not a Parent, not a Child, not a Sibbling
-                                #_father = item
-                                #_child  = x
-                                continue
-                        
-                            _relation = {}
-                            _relation['Father'] = _father
-                            _relation['Child'] = _child
-                            _relation["_lnkqty"] = int(reportLQI[item]['Neighbours'][x]['_lnkqty'], 16)
-                            _relation["DeviceType"] = reportLQI[item]['Neighbours'][x]['_devicetype']
-
-                            if _father != "0000":
-                                if 'ZDeviceName' in self.ListOfDevices[_father]:
-                                    if self.ListOfDevices[_father]['ZDeviceName'] != "" and self.ListOfDevices[_father]['ZDeviceName'] != {}:
-                                        #_relation[master] = self.ListOfDevices[_father]['ZDeviceName']
-                                        _relation['Father'] = self.ListOfDevices[_father]['ZDeviceName']
-                            else:
-                                _relation['Father'] = "Zigate"
-
-                            if _child != "0000":
-                                if 'ZDeviceName' in self.ListOfDevices[_child]:
-                                    if self.ListOfDevices[_child]['ZDeviceName'] != "" and self.ListOfDevices[_child]['ZDeviceName'] != {}:
-                                        #_relation[slave] = self.ListOfDevices[_child]['ZDeviceName']
-                                        _relation['Child'] = self.ListOfDevices[_child]['ZDeviceName']
-                            else:
-                                _relation['Child'] = "Zigate"
-
-                            # Sanity check, remove the direct loop
-                            if ( _relation['Child'], _relation['Father'] ) in _check_duplicate:
-                                self.logging( 'Debug', "Skip (%s,%s) as there is already ( %s, %s)" %(_relation['Father'], _relation['Child'], _relation['Child'], _relation['Father']))
-                                continue
-
-                            _check_duplicate.append( ( _relation['Father'], _relation['Child']))
-                            self.logging( 'Debug', "%10s Relationship - %15.15s - %15.15s %3s %2s" \
-                                %( _ts, _relation['Father'], _relation['Child'], _relation["_lnkqty"],
-                                        reportLQI[item]['Neighbours'][x]['_depth']))
-                            _topo[_ts].append( _relation )
-                        #end for x
-                    #end for item
-
-                    # Sanity check, to see if all devices are part of the report.
-                    # for iterDev in self.ListOfDevices:
-                    #     if iterDev in _nwkid_list: continue
-                    #     if 'Status' not in self.ListOfDevices[iterDev]: continue
-                    #     if self.ListOfDevices[iterDev]['Status'] != 'inDB': continue
-                    #    self.logging( 'Debug', "Nwkid %s has not been reported by this scan" %iterDev)
-                    #    _relation = {}
-                    #    _relation['Father'] = _relation['Child'] = iterDev
-                    #    _relation['_lnkqty'] = 0
-                    #    _relation['DeviceType'] = ''
-                    #    if 'ZDeviceName' in self.ListOfDevices[iterDev]:
-                    #        if self.ListOfDevices[iterDev]['ZDeviceName'] != "" and self.ListOfDevices[iterDev]['ZDeviceName'] != {}:
-                    #            _relation['Father'] = _relation['Child'] = self.ListOfDevices[iterDev]['ZDeviceName']
-                    #    _topo[_ts].append( _relation )
-
-                #end for _st
-
-        if verb == 'DELETE':
-            if len(parameters) == 0:
-                os.remove( _filename )
-                action = {}
-                action['Name'] = 'File-Removed'
-                action['FileName'] = _filename
-                _response['Data'] = json.dumps( action , sort_keys=True)
-
-            elif len(parameters) == 1:
-                timestamp = parameters[0]
-                if timestamp in _topo:
-                    self.logging( 'Debug', "Removing Report: %s from %s records" %(timestamp, len(_topo)))
-                    with open( _filename, 'r+') as handle:
-                        d = handle.readlines()
-                        handle.seek(0)
-                        for line in d:
-                            if line[0] != '{' and line[-1] != '}':
-                                handle.write( line )
-                                continue
-                            entry = json.loads( line, encoding=dict )
-                            entry_ts = entry.keys()
-                            if len( entry_ts ) == 1:
-                                if timestamp in entry_ts:
-                                    self.logging( 'Debug', "--------> Skiping %s" %timestamp)
-                                    continue
-                            else:
-                                continue
-                            handle.write( line )
-                        handle.truncate()
-
-                    action = {}
-                    action['Name'] = 'Report %s removed' %timestamp
-                    _response['Data'] = json.dumps( action , sort_keys=True)
-                else:
-                    Domoticz.Error("Removing Topo Report %s not found" %timestamp )
-                    _response['Data'] = json.dumps( [] , sort_keys=True)
-            return _response
-
-        if verb == 'GET':
-            if len(parameters) == 0:
-                # Send list of Time Stamps
-                _response['Data'] = json.dumps( _timestamps_lst , sort_keys=True)
-
-            elif len(parameters) == 1:
-                timestamp = parameters[0]
-                if timestamp in _topo:
-                    self.logging( 'Debug', "Topologie sent: %s" %_topo[timestamp])
-                    _response['Data'] = json.dumps( _topo[timestamp] , sort_keys=True)
-                else:
-                    _response['Data'] = json.dumps( [] , sort_keys=True)
-
         return _response
 
     def rest_nwk_stat( self, verb, data, parameters):
@@ -1142,121 +778,6 @@ class WebServer(object):
                 _response["Data"] = json.dumps( device_lst, sort_keys=True )
         return _response
 
-    def rest_zGroup_lst_avlble_dev( self, verb, data, parameters):
-
-        _response = setupHeadersResponse()
-        if self.pluginconf.pluginConf['enableKeepalive']:
-            _response["Headers"]["Connection"] = "Keep-alive"
-        else:
-            _response["Headers"]["Connection"] = "Close"
-        _response["Data"] = {}
-        _response["Status"] = "200 OK"
-        _response["Headers"]["Content-Type"] = "application/json; charset=utf-8"
-
-        if verb == 'GET':
-            device_lst = []
-            _device =  {}
-            _widget = {}
-            _device['_NwkId'] = '0000'
-            _device['WidgetList'] = []
-
-            _widget['_ID'] =  ''
-            _widget['Name'] =  ''
-            _widget['IEEE'] =  '0000000000000000'
-            _widget['Ep'] =  '01' 
-            _widget['ZDeviceName'] =  'Zigate (Coordinator)'
-            if self.zigatedata:
-                if 'IEEE' in self.zigatedata:
-                    _widget['IEEE'] =  self.zigatedata['IEEE'] 
-                    _device['_NwkId'] = self.zigatedata['Short Address']
-
-            _device['WidgetList'].append( _widget )
-            device_lst.append( _device )
-
-            for x in self.ListOfDevices:
-                if x == '0000':  
-                    continue
-
-                if 'MacCapa' not in self.ListOfDevices[x]:
-                    self.logging( 'Debug', "rest_zGroup_lst_avlble_dev - no 'MacCapa' info found for %s!!!!" %x)
-                    continue
-                
-                IkeaRemote = False
-                if 'Type' in self.ListOfDevices[x]:
-                    if self.ListOfDevices[x]['Type'] == 'Ikea_Round_5b':
-                        IkeaRemote = True
-
-                if self.ListOfDevices[x]['MacCapa'] != '8e' and not IkeaRemote:
-                    self.logging( 'Debug', "rest_zGroup_lst_avlble_dev - %s not a Main Powered device. " %x)
-                    continue
-
-                if 'Ep' in self.ListOfDevices[x]:
-                    if 'ZDeviceName' in self.ListOfDevices[x] and \
-                          'IEEE' in self.ListOfDevices[x]:
-                        _device = {}
-                        _device['_NwkId'] = x
-                        _device['WidgetList'] = []
-                        for ep in self.ListOfDevices[x]['Ep']:
-                            if 'Type' in self.ListOfDevices[x]:
-                                if self.ListOfDevices[x]['Type'] == 'Ikea_Round_5b':
-                                    if ep == '01':
-                                        if 'ClusterType' in self.ListOfDevices[x]['Ep']['01']:
-                                            widgetID = ''
-                                            for iterDev in self.ListOfDevices[x]['Ep']['01']['ClusterType']:
-                                                if self.ListOfDevices[x]['Ep']['01']['ClusterType'][iterDev] == 'Ikea_Round_5b':
-                                                    widgetID = iterDev
-                                                    for widget in self.Devices:
-                                                        if self.Devices[widget].ID == int(widgetID):
-                                                            _widget = {}
-                                                            _widget['_ID'] =  self.Devices[widget].ID 
-                                                            _widget['Name'] =  self.Devices[widget].Name 
-                                                            _widget['IEEE'] =  self.ListOfDevices[x]['IEEE'] 
-                                                            _widget['Ep'] =  ep 
-                                                            _widget['ZDeviceName'] =  self.ListOfDevices[x]['ZDeviceName'] 
-                                                            if _widget not in _device['WidgetList']:
-                                                                _device['WidgetList'].append( _widget )
-                                                            break
-                                                    if _device not in device_lst:
-                                                        device_lst.append( _device )
-                                            continue # Next Ep
-
-                            if '0004' not in self.ListOfDevices[x]['Ep'][ep] and \
-                                ( 'ClusterType' not in self.ListOfDevices[x]['Ep'][ep] or 'ClusterType' not in self.ListOfDevices[x]) and \
-                                '0006' not in self.ListOfDevices[x]['Ep'][ep] and \
-                                '0008' not in  self.ListOfDevices[x]['Ep'][ep] and \
-                                '0102' not in  self.ListOfDevices[x]['Ep'][ep]:
-                                continue
-
-                            if 'ClusterType' in self.ListOfDevices[x]['Ep'][ep] or 'ClusterType' in self.ListOfDevices[x]:
-                                if 'ClusterType' in self.ListOfDevices[x]['Ep'][ep]:
-                                    clusterType= self.ListOfDevices[x]['Ep'][ep]['ClusterType']
-                                else:
-                                    clusterType = self.ListOfDevices[x]['ClusterType']
-
-                                for widgetID in clusterType:
-                                    if clusterType[widgetID] not in ( 'LvlControl', 'Switch', 'Plug', 
-                                        "SwitchAQ2", "DSwitch", "Button", "DButton", 'LivoloSWL', 'LivoloSWR',
-                                        'ColorControlRGB', 'ColorControlWW', 'ColorControlRGBWW', 'ColorControlFull', 'ColorControl',
-                                        'VenetianInverted', 'Venetian', 'WindowCovering' ):
-                                        continue
-
-                                    for widget in self.Devices:
-                                        if self.Devices[widget].ID == int(widgetID):
-                                            _widget = {}
-                                            _widget['_ID'] =  self.Devices[widget].ID 
-                                            _widget['Name'] =  self.Devices[widget].Name 
-                                            _widget['IEEE'] =  self.ListOfDevices[x]['IEEE'] 
-                                            _widget['Ep'] =  ep 
-                                            _widget['ZDeviceName'] =  self.ListOfDevices[x]['ZDeviceName'] 
-                                            if _widget not in _device['WidgetList']:
-                                                _device['WidgetList'].append( _widget )
-
-                if _device not in device_lst:
-                    device_lst.append( _device )
-            self.logging( 'Debug', "Response: %s" %device_lst)
-            _response["Data"] = json.dumps( device_lst, sort_keys=True )
-            return _response
-
     def rest_zDevice_name( self, verb, data, parameters):
 
         _response = setupHeadersResponse()
@@ -1554,337 +1075,6 @@ class WebServer(object):
 
         return _response
 
-    def rest_zGroup( self, verb, data, parameters):
-
-        _response = setupHeadersResponse()
-        if self.pluginconf.pluginConf['enableKeepalive']:
-            _response["Headers"]["Connection"] = "Keep-alive"
-        else:
-            _response["Headers"]["Connection"] = "Close"
-        _response["Data"] = {}
-        _response["Status"] = "200 OK"
-        _response["Headers"]["Content-Type"] = "application/json; charset=utf-8"
-
-        self.logging( 'Debug', "rest_zGroup - ListOfGroups = %s" %str(self.groupmgt))
-
-        if verb == 'GET':
-            if self.groupmgt is None:
-                return _response
-            ListOfGroups = self.groupmgt.ListOfGroups
-            if ListOfGroups is None or len(ListOfGroups) == 0:
-                return _response
-
-            if len(parameters) == 0:
-                zgroup_lst = []
-                for itergrp in ListOfGroups:
-                    self.logging( 'Debug', "Process Group: %s" %itergrp)
-                    zgroup = {}
-                    zgroup['_GroupId'] = itergrp
-                    zgroup['GroupName'] = ListOfGroups[itergrp]['Name']
-                    zgroup['Devices'] = []
-                    for itemDevice in ListOfGroups[itergrp]['Devices']:
-                        if len(itemDevice) == 2:
-                            dev, ep = itemDevice
-                            ieee = self.ListOfDevices[dev]['IEEE']
-                        elif len(itemDevice) == 3:
-                            dev, ep, ieee = itemDevice
-                        self.logging( 'Debug', "--> add %s %s %s" %(dev, ep, ieee))
-                        _dev = {}
-                        _dev['_NwkId'] = dev
-                        _dev['Ep'] = ep
-                        _dev['IEEE'] = ieee
-                        zgroup['Devices'].append( _dev )
-
-                    if 'WidgetStyle' in ListOfGroups[itergrp]:
-                        zgroup['WidgetStyle'] = ListOfGroups[itergrp]['WidgetStyle']
-                    if 'Cluster' in ListOfGroups[itergrp]:
-                        zgroup['Cluster'] = ListOfGroups[itergrp]['Cluster']
-
-                    # Let's check if we don't have an Ikea Remote in the group
-                    if 'Tradfri Remote' in ListOfGroups[itergrp]:
-                        self.logging( 'Debug', "--> add Ikea Tradfri Remote")
-                        _dev = {}
-                        _dev['_NwkId'] = ListOfGroups[itergrp]["Tradfri Remote"]["Device Addr"]
-                        _dev['Ep'] = "01"
-                        zgroup['Devices'].append( _dev )
-                    zgroup_lst.append(zgroup)
-                self.logging( 'Debug', "zGroup: %s" %zgroup_lst)
-                _response["Data"] = json.dumps( zgroup_lst, sort_keys=True )
-
-            elif len(parameters) == 1:
-                if parameters[0] in ListOfGroups:
-                    itemGroup =  parameters[0]
-                    zgroup = {}
-                    zgroup['_GroupId'] = itemGroup
-                    zgroup['GroupName'] = ListOfGroups[itemGroup]['Name']
-                    zgroup['Devices'] = {}
-                    for itemDevice in ListOfGroups[itemGroup]['Devices']:
-                        if len(itemDevice) == 2:
-                            dev, ep = itemDevice
-                            _ieee = self.ListOfDevices[dev]['IEEE']
-                        elif len(itemDevice) == 3:
-                            dev, ep, _ieee = itemDevice
-                        self.logging( 'Debug', "--> add %s %s" %(dev, ep))
-                        zgroup['Devices'][dev] = ep 
-                    # Let's check if we don't have an Ikea Remote in the group
-                    if 'Tradfri Remote' in ListOfGroups[itemGroup]:
-                        self.logging( 'Log', "--> add Ikea Tradfri Remote")
-                        _dev = {}
-                        _dev['_NwkId'] = ListOfGroups[itemGroup]["Tradfri Remote"]["Device Addr"]
-                        _dev['Ep'] = "01"
-                        zgroup['Devices'].append( _dev )
-                    _response["Data"] = json.dumps( zgroup, sort_keys=True )
-
-        elif verb == 'PUT':
-            _response["Data"] = None
-            if  not self.groupmgt:
-                Domoticz.Error("Looks like Group Management is not enabled")
-                _response["Data"] = {}
-                return _response
-
-            ListOfGroups = self.groupmgt.ListOfGroups
-            grp_lst = []
-            if len(parameters) == 0:
-                self.restart_needed['RestartNeeded'] = True
-                data = data.decode('utf8')
-                data = json.loads(data)
-                self.logging( 'Debug', "data: %s" %data)
-                for item in data:
-                    self.logging( 'Debug', "item: %s" %item)
-                    if '_GroupId' not in item:
-                        self.logging( 'Debug', "--->Adding Group: ")
-                        # Define a GroupId 
-                        for x in range( 0x0001, 0x0999):
-                            grpid = '%04d' %x
-                            if grpid not in ListOfGroups:
-                                break
-                        else:
-                            Domoticz.Error("Out of GroupId")
-                            continue
-                        ListOfGroups[grpid] = {}
-                        ListOfGroups[grpid]['Name'] = item['GroupName']
-                        ListOfGroups[grpid]['Devices'] = []
-                    else:
-                        if item['_GroupId'] not in ListOfGroups:
-                            Domoticz.Error("zGroup REST API - unknown GroupId: %s" %grpid)
-                            continue
-                        grpid = item['_GroupId']
-
-                    grp_lst.append( grpid ) # To memmorize the list of Group
-                    #Update Group
-                    self.logging( 'Debug', "--->Checking Group: %s" %grpid)
-                    if item['GroupName'] != ListOfGroups[grpid]['Name']:
-                        # Update the GroupName
-                        self.logging( 'Debug', "------>Updating Group from :%s to : %s" %( ListOfGroups[grpid]['Name'], item['GroupName']))
-                        self.groupmgt._updateDomoGroupDeviceWidgetName( item['GroupName'], grpid )
-
-                    newdev = []
-                    if 'devicesSelected' not in item:
-                        continue
-                    if 'GroupName' in item:
-                        if item['GroupName'] == '':
-                            continue
-                    for devselected in item['devicesSelected']:
-                        if 'IEEE' in devselected:
-                            ieee = devselected['IEEE']
-                        elif '_NwkId' in devselected:
-                            nwkid = devselected['_NwkId']
-                            if nwkid != '0000' and nwkid not in self.ListOfDevices:
-                                Domoticz.Error("Not able to find nwkid: %s" %(nwkid))
-                                continue
-                            if 'IEEE' not in self.ListOfDevices[nwkid]:
-                                Domoticz.Error("Not able to find IEEE for %s - no IEEE entry in %s" %(nwkid, self.ListOfDevices[nwkid]))
-                                continue
-                            ieee = self.ListOfDevices[nwkid]['IEEE']
-                        else: 
-                            Domoticz.Error("Not able to find IEEE for %s" %(_dev))
-                            continue
-                        self.logging( 'Debug', "------>Checking device : %s/%s" %(devselected['_NwkId'], devselected['Ep']))
-                        # Check if this is not an Ikea Tradfri Remote
-                        nwkid = devselected['_NwkId']
-                        _tradfri_remote = False
-                        if 'Ep' in self.ListOfDevices[nwkid]:
-                            if '01' in self.ListOfDevices[nwkid]['Ep']:
-                                if 'ClusterType' in self.ListOfDevices[nwkid]['Ep']['01']:
-                                    for iterDev in self.ListOfDevices[nwkid]['Ep']['01']['ClusterType']:
-                                        if self.ListOfDevices[nwkid]['Ep']['01']['ClusterType'][iterDev] == 'Ikea_Round_5b':
-                                            # We should not process it through the group.
-                                            self.logging( 'Debug', "------>Not processing Ikea Tradfri as part of Group. Will enable the Left/Right actions")
-                                            ListOfGroups[grpid]['Tradfri Remote'] = {}
-                                            ListOfGroups[grpid]['Tradfri Remote']['Device Addr'] = nwkid
-                                            ListOfGroups[grpid]['Tradfri Remote']['Device Id'] = iterDev
-                                            _tradfri_remote = True
-                        if _tradfri_remote:
-                            continue
-                        # Process the rest
-                        for itemDevice in ListOfGroups[grpid]['Devices']:
-                            if len(itemDevice) == 2:
-                                _dev, _ep = itemDevice
-                                _ieee = self.ListOfDevices[dev]['IEEE']
-                            elif len(itemDevice) == 3:
-                                _dev, _ep, _ieee = itemDevice
-
-                            if _dev == devselected['_NwkId'] and _ep == devselected['Ep']:
-                                if (ieee, _ep) not in newdev:
-                                    self.logging( 'Debug', "------>--> %s to be added to group %s" %( (ieee, _ep), grpid))
-                                    newdev.append( (ieee, _ep) )
-                                else:
-                                    self.logging( 'Debug', "------>--> %s already in %s" %( (ieee, _ep), newdev))
-                                break
-                        else:
-                            if (ieee, devselected['Ep']) not in newdev:
-                                self.logging( 'Debug', "------>--> %s to be added to group %s" %( (ieee, devselected['Ep']), grpid))
-                                newdev.append( (ieee, devselected['Ep']) )
-                            else:
-                                self.logging( 'Debug', "------>--> %s already in %s" %( (_dev, _ep), newdev))
-                    # end for devselecte
-
-                    if 'coordinatorInside' in item:
-                        if item['coordinatorInside']:
-                            if 'IEEE' in self.zigatedata:
-                                ieee_zigate = self.zigatedata['IEEE']
-                                if ( ieee_zigate, '01') not in newdev:
-                                    self.logging( 'Debug', "------>--> %s to be added to group %s" %( (ieee_zigate, '01'), grpid))
-                                    newdev.append( (ieee_zigate, _ep) )
-
-                    self.logging( 'Debug', "--->Devices Added: %s" %newdev)
-                    ListOfGroups[grpid]['Imported'] = list( newdev )
-                    self.logging( 'Debug', "--->Grp: %s - tobe Imported: %s" %(grpid, ListOfGroups[grpid]['Imported']))
-
-                # end for item / next group
-                # Finaly , we need to check if AnyGroup have been removed !
-                self.logging( 'Debug', "Group to be removed")
-                Domoticz.Log("ListOfGroups: %s" %ListOfGroups)
-                for grpid in ListOfGroups:
-                    if grpid not in grp_lst:
-                        self.logging( 'Debug', "--->Group %s has to be removed" %grpid)
-                        if 'Imported' in ListOfGroups[grpid]:
-                            del ListOfGroups[grpid]['Imported']
-                        ListOfGroups[grpid]['Imported'] = []
-
-                self.logging( 'Debug', "Group to be worked out")
-                for grpid in ListOfGroups:
-                    self.logging( 'Debug', "Group: %s" %grpid)
-                    if 'Imported' in ListOfGroups[grpid]:
-                        for dev in ListOfGroups[grpid]['Imported']:
-                            self.logging( 'Debug', "---> %s to be imported" %str(dev))
-
-                self.groupmgt.write_jsonZigateGroupConfig()
-            # end if len()
-        # end if Verb=
-
-        return _response
-
-    def rest_bindLSTcluster( self, verb, data, parameters):
-        _response = setupHeadersResponse()
-        if self.pluginconf.pluginConf['enableKeepalive']:
-            _response["Headers"]["Connection"] = "Keep-alive"
-        else:
-            _response["Headers"]["Connection"] = "Close"
-        _response["Status"] = "200 OK"
-        _response["Headers"]["Content-Type"] = "application/json; charset=utf-8"
-
-        bindCluster = []
-        for key in self.ListOfDevices:
-            if key == '0000': 
-                continue
-
-            for ep in self.ListOfDevices[key]['Ep']:
-                for cluster in self.ListOfDevices[key]['Ep'][ep]:
-                    if cluster in ZCL_CLUSTERS_ACT:
-                        if cluster not in bindCluster:
-                            bindCluster.append( cluster )
-        _response["Data"] = json.dumps( bindCluster )
-        return _response
-
-    def rest_bindLSTdevice( self, verb, data, parameters):
-
-        _response = setupHeadersResponse()
-        if self.pluginconf.pluginConf['enableKeepalive']:
-            _response["Headers"]["Connection"] = "Keep-alive"
-        else:
-            _response["Headers"]["Connection"] = "Close"
-        _response["Status"] = "200 OK"
-        _response["Headers"]["Content-Type"] = "application/json; charset=utf-8"
-        if len(parameters) == 1:
-            listofdevices = []
-            clustertobind = parameters[0]
-            for key in self.ListOfDevices:
-                if key == '0000': 
-                    continue
-
-                for ep in self.ListOfDevices[key]['Ep']:
-                    if clustertobind in self.ListOfDevices[key]['Ep'][ep]:
-                        dev={}
-                        dev['IEEE'] = self.ListOfDevices[key]['IEEE']
-                        dev['NwkId'] = key
-                        dev['Ep'] = ep
-                        dev['ZDeviceName'] = self.ListOfDevices[key]['ZDeviceName']
-                        if dev not in listofdevices:
-                            listofdevices.append( dev )
-            _response["Data"] = json.dumps( listofdevices )
-        else:
-            Domoticz.Error("Must have 1 argument. %s" %parameters)
-        return _response
-
-    def rest_binding( self, verb, data, parameters):
-        _response = setupHeadersResponse()
-        if self.pluginconf.pluginConf['enableKeepalive']:
-            _response["Headers"]["Connection"] = "Keep-alive"
-        else:
-            _response["Headers"]["Connection"] = "Close"
-        _response["Status"] = "200 OK"
-        _response["Headers"]["Content-Type"] = "application/json; charset=utf-8"
-
-        if verb == 'PUT':
-            _response["Data"] = None
-            if len(parameters) == 0:
-                data = data.decode('utf8')
-                data = json.loads(data)
-
-                if 'sourceIeee' not in data and \
-                        'sourceEp' not in data and \
-                        'destIeee' not in data and \
-                        'destEp' not in data and \
-                        'cluster' not in data:
-                    Domoticz.Log("-----> uncomplet json %s" %data)
-                    _response["Data"] = json.dumps("uncomplet json %s" %data)
-                    return _response
-
-                self.logging( 'Log', "rest_binding - Source: %s/%s Dest: %s/%s Cluster: %s" %(data['sourceIeee'], data['sourceEp'], data['destIeee'], data['destEp'], data['cluster']))
-                webBind( self, data['sourceIeee'], data['sourceEp'], data['destIeee'], data['destEp'], data['cluster'] )
-                _response["Data"] = json.dumps( "Binding cluster %s between %s/%s and %s/%s" %(data['cluster'], data['sourceIeee'], data['sourceEp'], data['destIeee'], data['destEp']))
-                return _response
-
-    def rest_unbinding( self, verb, data, parameters):
-        _response = setupHeadersResponse()
-        if self.pluginconf.pluginConf['enableKeepalive']:
-            _response["Headers"]["Connection"] = "Keep-alive"
-        else:
-            _response["Headers"]["Connection"] = "Close"
-        _response["Status"] = "200 OK"
-        _response["Headers"]["Content-Type"] = "application/json; charset=utf-8"
-
-        if verb == 'PUT':
-            _response["Data"] = None
-            if len(parameters) == 0:
-                data = data.decode('utf8')
-                data = json.loads(data)
-
-                if 'sourceIeee' not in data and \
-                        'sourceEp' not in data and \
-                        'destIeee' not in data and \
-                        'destEp' not in data and \
-                        'cluster' not in data:
-                    Domoticz.Log("-----> uncomplet json %s" %data)
-                    _response["Data"] = json.dumps("uncomplet json %s" %data)
-                    return _response
-
-                self.logging( 'Log', "rest_unbinding - Source: %s/%s Dest: %s/%s Cluster: %s" %(data['sourceIeee'], data['sourceEp'], data['destIeee'], data['destEp'], data['cluster']))
-                webUnBind( self, data['sourceIeee'], data['sourceEp'], data['destIeee'], data['destEp'], data['cluster'] )
-                _response["Data"] = json.dumps( "Binding cluster %s between %s/%s and %s/%s" %(data['cluster'], data['sourceIeee'], data['sourceEp'], data['destIeee'], data['destEp']))
-                return _response
-
     def rest_raw_command( self, verb, data, parameters):
 
         Domoticz.Log("raw_command - %s %s" %(verb, data))
@@ -2093,262 +1283,3 @@ class WebServer(object):
 
                 _response["Data"] = json.dumps( dev_capabilities )
                 return _response
-
-    def rest_new_hrdwr( self, verb, data, parameters):
-
-        """
-        This is call to Enable/Disable a Provisioning process. As a toggle you will enable the provisioning or disable it
-        it will return either Enable or Disable
-        """
-
-        _response = setupHeadersResponse()
-        if self.pluginconf.pluginConf['enableKeepalive']:
-            _response["Headers"]["Connection"] = "Keep-alive"
-        else:
-            _response["Headers"]["Connection"] = "Close"
-        _response["Data"] = {}
-        _response["Status"] = "200 OK"
-        _response["Headers"]["Content-Type"] = "application/json; charset=utf-8"
-
-        if verb == 'GET':
-            data = {}
-            if len(parameters) != 1:
-                Domoticz.Error("rest_new_hrdwr - unexpected parameter %s " %parameters)
-                _response["Data"] = { "unexpected parameter %s " %parameters}
-                return _response
-
-            if parameters[0] not in ( 'enable', 'cancel', 'disable' ):
-                Domoticz.Error("rest_new_hrdwr - unexpected parameter %s " %parameters[0])
-                _response["Data"] = { "unexpected parameter %s " %parameters[0] }
-                return _response
-            if parameters[0] == 'enable':
-                Domoticz.Log("Enable Assisted pairing")
-                if len(self.DevicesInPairingMode):
-                    del self.DevicesInPairingMode
-                    self.DevicesInPairingMode = []
-                if not self.zigatedata:
-                    # Seems we are in None mode - Testing for ben
-                    self.fakeDevicesInPairingMode = 0
-
-                if self.permitTojoin['Duration'] != 255 and self.pluginparameters['Mode1'] != 'None':
-                    ZigatePermitToJoin(self, ( 4 * 60 ))
-
-                _response["Data"] = { "start pairing mode at %s " %int(time()) }
-                return _response
-
-            if parameters[0] in ( 'cancel', 'disable'):
-                Domoticz.Log("Disable Assisted pairing")
-                if len(self.DevicesInPairingMode):
-                    del self.DevicesInPairingMode
-                    self.DevicesInPairingMode = []
-
-                if not self.zigatedata:
-                    # Seems we are in None mode - Testing for ben
-                    self.fakeDevicesInPairingMode = 0
-
-                if self.permitTojoin['Duration'] != 255 and self.pluginparameters['Mode1'] != 'None':
-                    ZigatePermitToJoin(self, 0)
-
-                _response["Data"] = { "stop pairing mode at %s " %int(time()) }
-                return _response
-
-    def rest_rcv_nw_hrdwr( self, verb, data, parameters):
-
-        """
-        Will return a status on the provisioning process. Either Enable or Disable and in case there is a new device provisionned
-        during the period, it will return the information captured.
-        """
-
-        _response = setupHeadersResponse()
-        if self.pluginconf.pluginConf['enableKeepalive']:
-            _response["Headers"]["Connection"] = "Keep-alive"
-        else:
-            _response["Headers"]["Connection"] = "Close"
-        _response["Data"] = {}
-        _response["Status"] = "200 OK"
-        _response["Headers"]["Content-Type"] = "application/json; charset=utf-8"
-
-        if verb == 'GET':
-            data = {}
-            data['NewDevices'] = []
-
-            if not self.zigatedata:
-                # Seems we are in None mode - Testing for ben
-                if self.fakeDevicesInPairingMode in ( 0, 1):
-                    # Do nothing just wait the next pool
-                    self.fakeDevicesInPairingMode += 1
-                    _response["Data"] = json.dumps( data )
-                    return _response
-
-                if self.fakeDevicesInPairingMode in ( 2, 3 ):
-                    self.fakeDevicesInPairingMode += 1
-                    newdev = {}
-                    newdev['NwkId'] = list(self.ListOfDevices.keys())[0]
-                    data['NewDevices'].append( newdev )
-                    _response["Data"] = json.dumps( data )
-                    return _response
-
-                if self.fakeDevicesInPairingMode in ( 4, 5 ):
-                    self.fakeDevicesInPairingMode += 1
-                    newdev = {}
-                    newdev['NwkId'] = list(self.ListOfDevices.keys())[0]
-                    data['NewDevices'].append( newdev )
-                    newdev = {}
-                    newdev['NwkId'] = list(self.ListOfDevices.keys())[1]
-                    data['NewDevices'].append( newdev )
-                    _response["Data"] = json.dumps( data )
-                    return _response
-
-                if self.fakeDevicesInPairingMode in ( 6, 7 ):
-                    self.fakeDevicesInPairingMode += 1
-                    self.DevicesInPairingMode.append( list(self.ListOfDevices.keys())[0] )
-                    self.DevicesInPairingMode.append( list(self.ListOfDevices.keys())[1] )
-                    self.DevicesInPairingMode.append( list(self.ListOfDevices.keys())[2] )
-
-            Domoticz.Log("Assisted Pairing: Polling: %s" %str(self.DevicesInPairingMode))
-            if len(self.DevicesInPairingMode) == 0:
-                Domoticz.Log("--> Empty queue")
-                _response["Data"] = json.dumps( data )
-                return _response
-
-            listOfPairedDevices = list(self.DevicesInPairingMode)
-            _fake = 0
-            for nwkid in listOfPairedDevices:
-                if not self.zigatedata:
-                    _fake += 1
-                newdev = {}
-                newdev['NwkId'] = nwkid
-
-                Domoticz.Log("--> New device: %s" %nwkid)
-                if 'Status' not in self.ListOfDevices[ nwkid ]:
-                    Domoticz.Error("Something went wrong as the device seems not be created")
-                    data['NewDevices'].append( newdev )
-                    continue
-
-                if self.ListOfDevices[ nwkid ]['Status'] in ( '004d', '0045', '0043', '8045', '8043') or ( _fake == 1):
-                    # Pairing in progress, just return the Nwkid
-                    data['NewDevices'].append( newdev )
-                    continue
-
-                elif self.ListOfDevices[ nwkid ]['Status'] == 'UNKNOW' or ( _fake == 2):
-                    Domoticz.Log("--> UNKNOW , removed %s from List" %nwkid)
-                    self.DevicesInPairingMode.remove( nwkid )
-                    newdev['ProvisionStatus'] = 'Failed'
-                    newdev['ProvisionStatusDesc'] = 'Failed'
-
-                elif self.ListOfDevices[ nwkid ]['Status'] == 'inDB':
-                    Domoticz.Log("--> inDB , removed %s from List" %nwkid)
-                    self.DevicesInPairingMode.remove( nwkid )
-                    newdev['ProvisionStatus'] = 'inDB'
-                    newdev['ProvisionStatusDesc'] = 'inDB'
-                else:
-                    Domoticz.Log("--> Unexpected , removed %s from List" %nwkid)
-                    self.DevicesInPairingMode.remove( nwkid )
-                    newdev['ProvisionStatus'] = 'Unexpected'
-                    newdev['ProvisionStatusDesc'] = 'Unexpected'
-                    Domoticz.Error('Unexpected')
-                    continue
-
-                newdev['IEEE'] = 'Unknown'
-                if 'IEEE' in self.ListOfDevices[ nwkid ]:
-                    newdev['IEEE'] = self.ListOfDevices[ nwkid ]['IEEE']
-
-                newdev['ProfileId'] = ''
-                newdev['ProfileIdDesc'] = 'Unknow'
-                if 'ProfileID' in self.ListOfDevices[ nwkid ]:
-                    if self.ListOfDevices[ nwkid ]['ProfileID'] != {}:
-                        newdev['ProfileId'] = self.ListOfDevices[ nwkid ]['ProfileID']
-                        if int(newdev['ProfileId'],16) in PROFILE_ID:
-                            newdev['ProfileIdDesc'] = PROFILE_ID[ int(newdev['ProfileId'],16) ]
-
-                newdev['ZDeviceID'] = ''
-                newdev['ZDeviceIDDesc'] = 'Unknow'
-                if 'ZDeviceID' in self.ListOfDevices[ nwkid ]:
-                    if self.ListOfDevices[ nwkid ]['ZDeviceID'] != {}:
-                        newdev['ZDeviceID'] = self.ListOfDevices[ nwkid ]['ZDeviceID']
-                        if int(newdev['ProfileId'],16) == 0x0104: # ZHA
-                            if int(newdev['ZDeviceID'],16) in ZHA_DEVICES:
-                                newdev['ZDeviceIDDesc'] = ZHA_DEVICES[ int(newdev['ZDeviceID'],16) ]
-                            else:
-                                newdev['ZDeviceIDDesc'] = 'Unknow'
-                        elif int(newdev['ProfileId'],16) == 0xc05e: # ZLL
-                            if int(newdev['ZDeviceID'],16) in ZLL_DEVICES:
-                                newdev['ZDeviceIDDesc'] = ZLL_DEVICES[ int(newdev['ZDeviceID'],16) ]
-        
-                if 'Model' in self.ListOfDevices[ nwkid ]:
-                    newdev['Model'] = self.ListOfDevices[ nwkid ]['Model']
-    
-                newdev['PluginCertified'] = 'Unknow'
-                if 'ConfigSource' in self.ListOfDevices[nwkid]:
-                    if self.ListOfDevices[nwkid]['ConfigSource'] == 'DeviceConf':
-                        newdev['PluginCertified'] = 'yes'
-                    else:
-                        newdev['PluginCertified'] = 'no'
-    
-                newdev['Ep'] = []
-                if 'Ep' in self.ListOfDevices[ nwkid ]:
-                    for iterEp in  self.ListOfDevices[ nwkid ][ 'Ep' ]:
-                        ep = {}
-                        ep['Ep'] = iterEp
-                        ep['Clusters'] = []
-                        for clusterId in self.ListOfDevices[ nwkid ][ 'Ep' ][ iterEp ]:
-                            if clusterId in ( 'ClusterType', 'Type', 'ColorControl' ): 
-                                continue
-
-                            cluster = {}
-                            cluster['ClusterId'] = clusterId
-                            if clusterId in ZCL_CLUSTERS_LIST:
-                                cluster['ClusterDesc'] = ZCL_CLUSTERS_LIST[ clusterId ]
-                            else:
-                                cluster['ClusterDesc'] = 'Unknown'
-                            ep['Clusters'].append( cluster )
-                            Domoticz.Log("------> New Cluster: %s" %str(cluster))
-                        newdev['Ep'].append( ep )
-                        Domoticz.Log("----> New Ep: %s" %str(ep))
-                data['NewDevices'].append( newdev )
-                Domoticz.Log(" --> New Device: %s" %str(newdev))
-            # for nwkid in listOfPairedDevices:
-                    
-            _response["Data"] = json.dumps( data )
-            return _response
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
